requirements.txt
# Core AI Agent
# groq==0.4.2

# Google Calendar
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
google-api-python-client==2.172.0

# Deepgram (TTS/STT)
deepgram-sdk==2.12.0

# Server + WebSockets
fastapi==0.110.0
uvicorn==0.27.1
websockets==15.0.1
python-multipart==0.0.9
PyAudio==0.2.14
webrtcvad==2.0.10

# Utilities
pydantic==2.10.6
dateparser==1.2.0
python-dotenv==1.0.0
pytz==2025.2
python-dateutil==2.9.0.post0
pydub==0.25.1

# Google Generative AI
google-generativeai
google-ai-generativelanguage

aenum==3.1.15
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
boto3==1.34.34
botocore==1.34.162
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.2.1
colorama==0.4.6
dataclasses-json==0.6.7
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
dotenv==0.9.9
filelock==3.18.0
Flask==3.1.1
frozenlist==1.5.0
fsspec==2025.3.2
future==1.0.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.25.1
googleapis-common-protos==1.70.0
greenlet==3.1.1
grpcio==1.73.0
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.8
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
langchain
langchain-core
langchain-google-genai
langchain-groq==0.3.2
langchain-text-splitters==0.3.8
langsmith==0.3.45
llvmlite==0.44.0
loguru==0.7.3
Markdown==3.8
MarkupSafe==3.0.2
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.4.3
mypy-extensions==1.0.0
narwhals==1.42.1
networkx==3.4.2
numba==0.61.2
numpy==1.26.4
oauthlib==3.2.2
openai==1.70.0
orjson==3.10.16
ormsgpack==1.10.0
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.3
pillow==11.1.0
pipecat-ai @ git+https://github.com/pipecat-ai/pipecat.git@6278561f888db6ea3777d8f1ff53497f95461630
propcache==0.3.1
proto-plus==1.26.1
protobuf==4.25.8
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycparser==2.22
pydantic_core==2.27.2
pydeck==0.9.1
pyloudnorm==0.1.1
pymongo==4.6.1
pyparsing==3.2.3
PySocks==1.7.1
PyWavefront==1.3.3
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
resampy==0.4.3
rpds-py==0.25.1
rsa==4.9.1
s3transfer==0.10.4
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
selenium==4.32.0
sentence-transformers==4.0.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soxr==0.5.0.post1
SpeechRecognition==3.14.3
SQLAlchemy==2.0.40
streamlit==1.45.1
sympy==1.13.1
tenacity==8.5.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
torch==2.6.0
tornado==6.5.1
tqdm==4.67.1
transformers==4.51.2
trio==0.30.0
trio-websocket==0.12.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
vaderSentiment==3.3.2
watchdog==6.0.0
webrtcvad==2.0.10
websocket-client==1.8.0
Werkzeug==3.1.3
win32_setctime==1.2.0
wsproto==1.2.0
xxhash==3.5.0
zstandard==0.23.0

